requirements.txt
kafka-python
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
dbt-core
dbt-postgres
fastapi
uvicorn